# api/openapi/cbizdocsmanager-v1.yaml
openapi: 3.0.3
info:
  title: CBizDocsManager API
  version: 1.0.0
  description: 公务单据管理系统后端 API v1 契约。
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          description: 用户名或密码错误（AUTH_INVALID_CREDENTIALS）
          x-error-codes: [AUTH_INVALID_CREDENTIALS]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: 使用邀请码注册组内子账号
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterApiResponse'
        '400':
          description: 请求或邀请码无效（VALIDATION_FAILED、INVITATION_INVALID、INVITATION_EXPIRED、INVITATION_USED）
          x-error-codes:
            - VALIDATION_FAILED
            - INVITATION_INVALID
            - INVITATION_EXPIRED
            - INVITATION_USED
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 用户名已存在（USER_USERNAME_EXISTS）
          x-error-codes: [USER_USERNAME_EXISTS]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: 刷新访问令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 刷新成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          description: 刷新令牌无效或已过期（AUTH_REFRESH_INVALID、AUTH_TOKEN_EXPIRED）
          x-error-codes: [AUTH_REFRESH_INVALID, AUTH_TOKEN_EXPIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: 注销当前会话
      responses:
        '200':
          description: 注销成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: 获取当前用户与所属组摘要
      responses:
        '200':
          description: 查询成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/password:
    put:
      tags: [Auth]
      operationId: changePassword
      summary: 修改当前用户密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 密码修改成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordChangedApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          description: 当前密码错误或访问令牌已过期（AUTH_INVALID_CREDENTIALS、AUTH_TOKEN_EXPIRED）
          x-error-codes: [AUTH_INVALID_CREDENTIALS, AUTH_TOKEN_EXPIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/platform/groups:
    post:
      tags: [Platform]
      operationId: createGroup
      summary: 平台管理员创建组及主账号
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: 组与主账号创建成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGroupApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 组名或主账号用户名已存在（GROUP_NAME_EXISTS、USER_USERNAME_EXISTS）
          x-error-codes: [GROUP_NAME_EXISTS, USER_USERNAME_EXISTS]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Platform]
      operationId: listGroups
      summary: 平台管理员分页查询业务组
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/GroupStatus'
        - name: keyword
          in: query
          required: false
          description: 按组名模糊匹配
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPageApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/platform/groups/{group_id}:
    get:
      tags: [Platform]
      operationId: getGroupDetail
      summary: 平台管理员查询组治理详情
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      responses:
        '200':
          description: 查询成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetailApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: 组不存在（GROUP_NOT_FOUND）
          x-error-codes: [GROUP_NOT_FOUND]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/platform/groups/{group_id}/status:
    patch:
      tags: [Platform]
      operationId: changeGroupStatus
      summary: 平台管理员停用或启用业务组
      description: 目标状态与当前状态一致时按幂等成功处理；停用组会连同撤销该组全部刷新会话。
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeGroupStatusRequest'
      responses:
        '200':
          description: 变更成功，返回组最新摘要
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupSummaryApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: 组不存在（GROUP_NOT_FOUND）
          x-error-codes: [GROUP_NOT_FOUND]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 组已被其他请求修改（RESOURCE_VERSION_CONFLICT）
          x-error-codes: [RESOURCE_VERSION_CONFLICT]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/platform/groups/{group_id}/owner:
    put:
      tags: [Platform]
      operationId: changeGroupOwner
      summary: 平台管理员交接组主账号
      description: 支持把现有成员提升为主账号（existing_member）或新建账号并设为主账号（new_account）。
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeGroupOwnerRequest'
      responses:
        '200':
          description: 交接成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerChangedApiResponse'
        '400':
          description: 请求参数错误或交接目标无效（VALIDATION_FAILED、OWNER_TARGET_INVALID）
          x-error-codes: [VALIDATION_FAILED, OWNER_TARGET_INVALID]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: 组不存在（GROUP_NOT_FOUND）
          x-error-codes: [GROUP_NOT_FOUND]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 组已被其他请求修改或用户名已存在（RESOURCE_VERSION_CONFLICT、USER_USERNAME_EXISTS）
          x-error-codes: [RESOURCE_VERSION_CONFLICT, USER_USERNAME_EXISTS]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/groups/invitations:
    post:
      tags: [Group]
      operationId: createInvitation
      summary: 组主账号创建子账号邀请码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: 邀请码创建成功；明文邀请码仅在本响应返回
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInvitationApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无组管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Group]
      operationId: listInvitations
      summary: 组主账号分页查询邀请码列表
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/InvitationDisplayStatus'
      responses:
        '200':
          description: 查询成功；列表永不返回邀请码明文与密文
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationPageApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无组管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/groups/invitations/{invitation_id}/secret:
    post:
      tags: [Group]
      operationId: revealInvitation
      summary: 组主账号查看仍有效的邀请码明文
      description: 仅 active 且未过期的邀请码可查看；响应带 no-store，客户端展示后必须立即清除本地副本。
      parameters:
        - $ref: '#/components/parameters/InvitationIdPath'
      responses:
        '200':
          description: 查看成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Cache-Control:
              description: 固定为 no-store，禁止代理与浏览器缓存明文邀请码
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationSecretApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无组管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 邀请码已使用、已撤销或已过期，不再可查看（INVITATION_NOT_REVEALABLE）
          x-error-codes: [INVITATION_NOT_REVEALABLE]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          description: 密钥缺失或密文校验失败（INTERNAL_ERROR、INVITATION_DECRYPT_FAILED）
          x-error-codes: [INTERNAL_ERROR, INVITATION_DECRYPT_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
  /api/v1/groups/invitations/{invitation_id}/revoke:
    post:
      tags: [Group]
      operationId: revokeInvitation
      summary: 组主账号撤销仍有效的邀请码
      parameters:
        - $ref: '#/components/parameters/InvitationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeInvitationRequest'
      responses:
        '200':
          description: 撤销成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationSummaryApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无组管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: 邀请码不存在（INVITATION_NOT_FOUND）
          x-error-codes: [INVITATION_NOT_FOUND]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 邀请码不可撤销或已被其他请求修改（INVITATION_NOT_REVOKABLE、RESOURCE_VERSION_CONFLICT）
          x-error-codes: [INVITATION_NOT_REVOKABLE, RESOURCE_VERSION_CONFLICT]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/web/login:
    post:
      tags: [Auth]
      operationId: webLogin
      summary: Web 用户登录并设置 Refresh 与 CSRF Cookie
      security: []
      parameters:
        - $ref: '#/components/parameters/OriginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功；响应体不包含 Refresh Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebAuthTokenApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          description: 用户名或密码错误（AUTH_INVALID_CREDENTIALS）
          x-error-codes: [AUTH_INVALID_CREDENTIALS]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '403':
          description: 请求来源不允许（ORIGIN_FORBIDDEN）
          x-error-codes: [ORIGIN_FORBIDDEN]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/auth/web/logout:
    post:
      tags: [Auth]
      operationId: webLogout
      summary: 注销 Web Cookie 会话
      description: Refresh Token 只从 HttpOnly Cookie 读取，不接收 JSON Refresh Token。
      security: []
      parameters:
        - $ref: '#/components/parameters/OriginHeader'
        - $ref: '#/components/parameters/CSRFHeader'
      responses:
        '200':
          description: 注销成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutApiResponse'
        '401':
          description: Refresh Cookie 无效（AUTH_REFRESH_INVALID）
          x-error-codes: [AUTH_REFRESH_INVALID]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '403':
          description: CSRF 或 Origin 校验失败（CSRF_INVALID、ORIGIN_FORBIDDEN）
          x-error-codes: [CSRF_INVALID, ORIGIN_FORBIDDEN]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/groups/members:
    get:
      tags: [Group]
      operationId: listGroupMembers
      summary: 分页查询当前组成员
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled, removed]
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPageApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无成员管理权限（FORBIDDEN）
          x-error-codes: [FORBIDDEN]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestID:
      description: 本次请求的唯一追踪 ID
      schema:
        type: string
...
  responses:
...
    InternalError:
      description: 服务内部错误（INTERNAL_ERROR）
      x-error-codes: [INTERNAL_ERROR]
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
  schemas:
...
    ErrorApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required: [code, message, data, request_id]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            data:
              type: object
              nullable: true
            field_errors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          format: password
          minLength: 1
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [invitation_code, username, password, display_name]
      properties:
        invitation_code:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 1
        password:
          type: string
          format: password
          minLength: 8
        display_name:
          type: string
          minLength: 1
...
    AuthTokenApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TokenData'
...
    GroupStatus:
      type: string
      enum: [active, disabled]
...